Basic Anti-Bot Protection
Simple coloured shapes have been used that shuffle and require a new user to select the "green circle" this allows those that don't speak english or happen to be colour blind to complete the entry test. If a user fails they are simply restricted from posting which should reduce the issues with accidental failures as they can contact an admin they see in the chat.

Was unable to find a way to time delete the messages, but showing the link preview on the message has been disabled to avoid it being too large (suggestions on how to remove the messages after a set amount of time would be great if you have any ideas).

Welcome message has also been adjusted to avoid rollover on mobile.