Detect & Translate Non-English Messages
Every non-command message is checked to see if it is likely to be English or not, if there is a high confident it is not English it will be translated into English. The library used to perform the language checks (langid) is a smaller standalone library so shouldn't be too heavy to use this way.